Separated quantux_libraries.py module
In the notebook load the quantux_libraries. Notice to use variables inside the quantux_libraries you have to use it as prefix like quantux_libraries._SEED to use the seed variable. The functions are loaded with the same names to the main module. Fuctions from quantum.libraries can be used straight.

Added optuna comparison to override the optuna dictionary for different kind of comparisons.